Moved to YAML based rules for rewriting
I think these might be a bit easier to maintain
