+++ requirements.in
@@ -7,3 +7,4 @@ requests
 newrelic
 h_pyramid_sentry
 whitenoise
+pyyaml
