+++ requirements.txt
@@ -21,6 +21,7 @@ plaster-pastedeploy==0.7  # via pyramid
 plaster==1.0              # via plaster-pastedeploy, pyramid
 pyramid-jinja2==2.8       # via -r requirements.in
 pyramid==1.10.4           # via -r requirements.in, h-pyramid-sentry, pyramid-jinja2
+pyyaml==5.3.1             # via -r requirements.in
 requests==2.23.0          # via -r requirements.in
 sentry-sdk==0.14.4        # via h-pyramid-sentry
 soupsieve==2.0            # via beautifulsoup4
