+++ via/services/rewriter/rewriter_html_htmlparser.py
@@ -3,7 +3,7 @@ from html.parser import HTMLParser
 from io import StringIO
 
 from via.services.rewriter.rewriter import AbstractHTMLRewriter
-from via.services.rewriter.rules import Attribute
+from via.services.rewriter.ruleset import Attribute
 from via.services.timeit import timeit
